added transform tests, cleaned up keyboard def validator tests
